fix: resolve issues with notifications
